config/supplychain/quorab/solvers/annealer_solver.py
import time
from typing import Any, Dict

import numpy as np


class AnnealerSolver:
    def __init__(self):
        import neal

        self.sampler = neal.SimulatedAnnealingSampler()

    def solve(self, Q: np.ndarray, num_reads: int = 1000) -> Dict[str, Any]:
        """
        Solves a QUBO problem using Simulated Annealing (proxy for D-Wave Annealer).
        Accepts a numpy matrix.
        """
        start_time = time.time()

        n = Q.shape[0]
        qubo_dict = {}
        for i in range(n):
            for j in range(i, n):
                if Q[i, j] != 0:
                    qubo_dict[(i, j)] = float(Q[i, j])

        import dimod

        bqm = dimod.BinaryQuadraticModel.from_qubo(qubo_dict)
        sampleset = self.sampler.sample(bqm, num_reads=num_reads)

        best_sample = sampleset.first.sample
        best_energy = sampleset.first.energy

        solve_time = time.time() - start_time

        # Convert bitstring sample to a standard format
        solution_dict = {f"x_{k}": v for k, v in best_sample.items()}

        return {
            "solution": solution_dict,
            "cost": best_energy,
            "solver_used": "Simulated Annealing (neal)",
            "performance_metrics": {"time_taken": solve_time, "iterations": num_reads},
        }
