config/supplychain/quorab/requirements.txt
fastapi>=0.103.0
uvicorn>=0.23.2
pydantic>=2.4.2
networkx>=3.1
numpy>=1.26.0
qiskit>=0.44.1
qiskit-optimization>=0.6.0
qiskit-aer>=0.12.2
dimod>=0.12.11
dwave-neal>=0.6.0
ortools>=9.7.2996
